优化界面的切换逻辑

+++ app/src/main/java/com/example/ttymyday/model/ScheduleTag.kt
@@ -1,6 +1,6 @@
 package com.example.ttymyday.model
 
-data class ScheduleTag(var id:Int,
+data class ScheduleTag(var id:Long,
                        var name:String?,
                        var title:String?,
                        var owner:String?,
